requirements.txt
# 10. requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
cellpose>=2.0.0
matplotlib>=3.3.0
pyyaml>=5.4.0
tqdm>=4.50.0
